Compute Integrated Baseline Hazard and Its Derivatives

This function computes the integrated baseline hazard or its derivatives using Gaussian quadrature and Bernstein polynomial basis functions.

Usage

int_basehaz(t, deriv, alpha, n_basis, boundary, n_nodes = 10)

Arguments

t

Numeric. The upper limit of integration.

deriv

Integer. Specifies the derivative to compute:

0: Computes the cumulative baseline hazard.

1: Computes the gradient of the cumulative baseline hazard.

2: Computes the Hessian of the cumulative baseline hazard.

alpha

Numeric vector. Coefficients for the baseline hazard function.

n_basis

Integer. The number of Bernstein polynomial basis functions.

boundary

Numeric vector of length 2. The boundary knots for the Bernstein polynomials.

n_nodes

Integer. The number of quadrature nodes to use. Default is 10.

Value

Numeric. The computed integral or its derivative.

Details

This function uses Gaussian quadrature to approximate the integral of the baseline hazard function, which is modeled using Bernstein polynomial basis functions. The deriv parameter determines whether the function computes the integral itself or its first or second derivative.